Memory Window: rewrite internals to better handle UI/state mismatch.
Now all the "UI state" is calculated together so it is consistent. This
includes the number of rows, column layout etc.

Then "RecalcText" is called to fill in all the characters for this layout,
but also handling memory being missing (when it shows "?")

When memory arrives, RecalcText() is called again to update with the new
memory contents.